Moving JetpackLogo component from Jetpack to RNA Components.
