+++ projects/js-packages/components/components/jetpack-logo/README.md
@@ -1,0 +1,18 @@
+JetpackLogo
+========
+
+Component that renders the Jetpack SVG logo.
+It consists of the Jetpack symbol followed by the name.
+It takes width and height properties but defaults to 32px in height.
+
+#### How to use:
+
+```js
+<JetpackLogo height={ 48 } className="jp-logo" />
+```
+
+#### Props
+
+* `className`: String - (default: `jetpack-logo`) the class name set on the SVG element.
+* `height`: Number - (default: 32) set the height of the logo.
+* `width`: Number - (optional) set the width of the logo.
