+++ projects/js-packages/components/index.jsx
@@ -13,3 +13,5 @@ You should have received a copy of the GNU General Public License
 along with this program; if not, write to the Free Software
 Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301, USA.
 */
+
+export { default as JetpackLogo } from './components/jetpack-logo';
