+++ projects/packages/connection-ui/_inc/components/header/index.jsx
@@ -3,6 +3,7 @@
  */
 import React from 'react';
 import { __ } from '@wordpress/i18n';
+import { JetpackLogo } from '@automattic/jetpack-components';
 
 /**
  * Internal dependencies
@@ -17,6 +18,7 @@ import './style.scss';
 const Header = () => {
 	return (
 		<div className="jetpack-cui__header">
+			<JetpackLogo />
 			<h1>{ __( 'Connection Manager', 'jetpack' ) }</h1>
 		</div>
 	);
