+++ projects/plugins/jetpack/_inc/client/components/jetpack-logo/README.md
@@ -1,18 +1,5 @@
 JetpackLogo
 ========
 
-Component that renders the Jetpack SVG logo.
-It consists of the Jetpack symbol followed by the name.
-It takes width and height properties but defaults to 32px in height.
-
-#### How to use:
-
-```js
-<JetpackLogo height={ 48 } className="jp-logo" />
-```
-
-#### Props
-
-* `className`: String - (default: `jetpack-logo`) the class name set on the SVG element.
-* `height`: Number - (default: 32) set the height of the logo.
-* `width`: Number - (optional) set the width of the logo.
+The component is deprecated.
+Use `@automattic/jetpack-components` instead.
