+++ projects/plugins/jetpack/package.json
@@ -59,6 +59,7 @@
 		"@automattic/calypso-color-schemes": "2.1.1",
 		"@automattic/components": "1.0.0-alpha.3",
 		"@automattic/format-currency": "1.0.0-alpha.0",
+		"@automattic/jetpack-components": "workspace:^0.1.0-alpha",
 		"@automattic/jetpack-connection": "workspace:^0.4.0-alpha",
 		"@automattic/popup-monitor": "1.0.0",
 		"@automattic/request-external-access": "1.0.0",
